requirements.txt
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.1
comtypes==1.1.4
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
filelock==3.13.4
fonttools==4.51.0
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
intel-openmp==2021.4.0
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
labelImg==1.8.6
lxml==5.2.1
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
mkl==2021.4.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.2
parso==0.8.3
pillow==10.3.0
pillow_heif==0.16.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
py-cpuinfo==9.0.0
Pygments==2.17.2
pykinect2==0.1.0
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
scikit-learn==1.5.0
scipy==1.13.0
seaborn==0.13.2
setuptools==68.2.2
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
sympy==1.12
tbb==2021.12.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
torch==2.3.0+cu121
torchaudio==2.3.0+cu121
torchvision==0.18.0+cu121
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.2.11
urllib3==2.2.1
wcwidth==0.2.13
wheel==0.41.2
